(Test_Task): changed type of member handled_ from int to size_t to avoid signed/unsigned comparison in Test_Task::handle_input ().

+++ trunk/tests/Reactors_Test.cpp
@@ -47,7 +47,7 @@ public:
                             ACE_Reactor_Mask close_mask);
 
 private:
-  int handled_;
+  size_t handled_;
   // Number of iterations handled.
 
   static int task_count_;
